backend/requirements.txt
tornado==6.4.1
pymongo==4.15.4
google-genai==0.3.0
pydantic==2.10.5
python-jose[cryptography]==3.3.0
pytest==8.3.4
pytest-asyncio==0.24.0
uvicorn==0.32.0
